fix: fix classify_input.py

+++ code-contest-exp/scripts/cgig/classify_input.py
@@ -2,15 +2,24 @@ import json
 import os
 from pathlib import Path
 from typing import List, Tuple, Dict, Union
-from multiprocessing import Pool, Manager
+from multiprocessing import Pool
 from tqdm import tqdm
 import tempdir
 import subprocess
 import shutil
 from fire import Fire
+import time
+import threading
 
 from config import config
 
+
+def clear_shared_memory():
+    """Continuously clears unused shared memory every 10 seconds."""
+    while True:
+        time.sleep(10)
+        print("Clearing shared memory at ", time.time())
+        subprocess.run("ipcs -m | awk '$6==0 {print $2}' | xargs -r -n 1 ipcrm -m", shell=True)
 
 def parse_stack_trace(stack_trace: str) -> Tuple[str, int]:
     """
@@ -47,7 +56,7 @@ def compile_solution(tmp_dir: Path, instrumented_program: Path) -> Union[Path, s
     return "Compile Error"
 
 
-def run_classifier(input_file_list: List[Path], instrumented_program: Path) -> Dict[str, Dict]:
+def run_classifier(input_file_list: List[Path], instrumented_program: Path, clear: bool = True) -> Dict[str, Dict]:
     """
     Run the classifier on the given input file list and monitor whether there is a crash,
     and use the stack trace to obtain the corresponding checking constraint and line number.
@@ -93,10 +102,24 @@ def run_classifier(input_file_list: List[Path], instrumented_program: Path) -> D
                     continue
                 finally:
                     # clear shared memory
-                    os.system("ipcs -m | awk '$6==0 {print $2}' | xargs -r -n 1 ipcrm -m")
+                    if clear:
+                        os.system("ipcs -m | awk '$6==0 {print $2}' | xargs -r -n 1 ipcrm -m")
 
     return res_dict
 
+def filter_input_files_for_alphacode(problem_id:str, input_file_list: List[Path]) -> List[Path]:
+    """
+    Filter the input files for alphacode.
+    the logic is slightly different from sanitize_alphacode_result.py
+    when no validator is available, here we just return the input_file_list
+    """
+    input_validation_file = Path(config["result_root_dir"]) / "input_validation_result" / "alphacode" / f"{problem_id}.json"
+    if not input_validation_file.exists():
+        # no available validator
+        return input_file_list
+    input_validation_result_dict = json.loads(input_validation_file.read_text())
+
+    return [f for f in input_file_list if input_validation_result_dict[f.name] == "PASS"]
 
 def process_strategy(
     strategy: str,
@@ -109,18 +132,22 @@ def process_strategy(
     Process a single strategy for a specific problem ID.
     """
     input_dir = problem_root_dir / problem_id / strategy / "input"
-    if strategy == "alphacode":
+    if strategy == "alphacode" or strategy == "alphacode_sanitized":
         input_dir = problem_root_dir / problem_id / "input"
 
     input_file_list = list(input_dir.glob("*.in"))
+    if strategy == "alphacode_sanitized":
+        input_file_list = filter_input_files_for_alphacode(problem_id, input_file_list)
 
     for solution_id in os.listdir(constraints_dir / problem_id):
         for input_pair in os.listdir(constraints_dir / problem_id / solution_id):
+            result_file = input_classify_dir / problem_id / solution_id / input_pair / f"{strategy}.json"
+            if result_file.exists():
+                continue
             instrumented_program = constraints_dir / problem_id / solution_id / input_pair / "transformed_program.cpp"
 
             assert instrumented_program.exists(), f"{instrumented_program} does not exist"
-            res = run_classifier(input_file_list, instrumented_program)
-            result_file = input_classify_dir / problem_id / solution_id / input_pair / f"{strategy}.json"
+            res = run_classifier(input_file_list, instrumented_program, clear=False)
             result_file.parent.mkdir(parents=True, exist_ok=True)
             with open(result_file, "w") as f:
                 json.dump(res, f, indent=4)
@@ -130,7 +157,10 @@ def main(problem_root_dir: str = config["problem_root_dir"]):
     constraints_dir = Path(config["constraints_dir"])
     problem_root_dir = Path(problem_root_dir)
     input_classify_dir = Path(config["input_classify_dir"])
-    strategies = ["alphacode", "plain_problem", "slow_solution", "diff_solution_one_input", "random_solution", "evalperf_slow_solution", "evalperf_random_solution", "corpus"]
+    strategies = ["alphacode_sanitized", "plain_problem", "evalperf_slow_solution", "evalperf_random_solution", "corpus_instrument_fuzz_mutator_with_constraint_per_solution", "corpus_raw_fuzz_custom_mutator", "corpus_raw_fuzz_default_mutator", "corpus_raw_fuzz_mutator_with_constraint_per_solution"]
+
+    cleanup_thread = threading.Thread(target=clear_shared_memory, daemon=True)
+    cleanup_thread.start()
 
     tasks = []
     for strategy in strategies:
